CLN: Abstract index functionality to parent class
